refactor: split SettingsContext into focused hooks
Closes #230
